Change integers to unsigned like in lodepng picture

+++ src/utilities/utils.cpp
@@ -101,13 +101,13 @@ int printPlatformProfile(bool print_extras)
   }
 
 
-void movingAvgFilter(unsigned char* image, int width, int height, int windowSize) {
+void movingAvgFilter(unsigned char* image, unsigned width, unsigned height, unsigned windowSize) {
   unsigned char* imageCopy = (unsigned char*)malloc(sizeof(unsigned char) * width * height);
   int windowSizeHalf = windowSize / 2;
 
   // Loop through the image
-  for (int y = 0; y < height; y++) {
-    for (int x = 0; x < width; x++) {
+  for (unsigned y = 0; y < height; y++) {
+    for (unsigned x = 0; x < width; x++) {
       int sum = 0;
       int count = 0;
 
@@ -118,7 +118,7 @@ void movingAvgFilter(unsigned char* image, int width, int height, int windowSize
           int y2 = y + i;
 
           // Check that the pixel is inside the image
-          if (x2 >= 0 && x2 < width && y2 >= 0 && y2 < height) {
+          if (x2 >= 0 && x2 < (int)width && y2 >= 0 && y2 < (int)height) {
             sum += image[y2 * width + x2];
             count++;
           }
